pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-trust-telemetry"
version = "0.1.0"
description = "Trust telemetry middleware for inter-agent communication — makes instruction contamination observable across traces"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "agent-trust-telemetry contributors" },
]
keywords = ["agent", "trust", "telemetry", "security", "opentelemetry", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
att = "att.cli:main"

[project.urls]
Homepage = "https://github.com/wharfe/agent-trust-telemetry"
Repository = "https://github.com/wharfe/agent-trust-telemetry"
Issues = "https://github.com/wharfe/agent-trust-telemetry/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/att"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = "src"
